backend/pyproject.toml
[project]
name = "finapp-backend"
version = "0.1.0"
description = "Personal finance dashboard backend"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "sqlalchemy>=2.0",
    "pydantic>=2.5",
    "pypdf>=4.0",
    "python-multipart>=0.0.9",
    "pydantic-settings>=2.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.3", "httpx>=0.27"]
azure = ["azure-ai-documentintelligence>=1.0.0"]
postgres = ["psycopg[binary]>=3.1", "alembic>=1.13"]
prod = [
    "azure-ai-documentintelligence>=1.0.0",
    "psycopg[binary]>=3.1",
    "alembic>=1.13",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
